% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TileScheme.R
\name{$,tileScheme-method}
\alias{$,tileScheme-method}
\title{Get and set data}
\usage{
\S4method{$}{tileScheme}(x, name)
}
\arguments{
\item{x}{a 'tileScheme' object}

\item{name}{name of data column}

\item{value}{vector of new data values}
}
\description{
Get and set data
}
