% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tileScheme-methods.R
\name{plot}
\alias{plot}
\alias{plot,tileScheme,ANY-method}
\title{Plot}
\usage{
plot(x, y, ...)

\S4method{plot}{tileScheme,ANY}(x, labels = TRUE, add = FALSE, ...)
}
\arguments{
\item{x}{a 'tileScheme' object}

\item{y}{argument not used for 'tileScheme' objects}

\item{...}{arguments t be passed to methods}

\item{labels}{logical. Add row and column labels}

\item{add}{logical. Plot 'tileScheme' on top of existing plot}
}
\description{
Plot a 'tileScheme' object
}
