% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tileScheme.R
\name{show,tileScheme-method}
\alias{show,tileScheme-method}
\title{Show}
\usage{
\S4method{show}{tileScheme}(object)
}
\arguments{
\item{object}{a 'tileScheme' object}
}
\description{
Print information about a 'tileScheme' object
}
