% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tileScheme-methods.R
\docType{methods}
\name{[}
\alias{[}
\title{Subset}
\usage{
\S4method{[}{tileScheme,ANY,ANY}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{x}{a 'tileScheme' object}

\item{i, j, ...}{indices specifying elements to extract}

\item{drop}{argument not used for 'tileScheme'}
}
\description{
Subset tiles using the single bracket operator. Subset geometry (tiles, buffs or nbuffs) using the double brackets
}
