% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TileHandlers.R
\name{TempTiles}
\alias{TempTiles}
\title{Generate Temporary Tiles}
\usage{
TempTiles(inRaster, tiles)
}
\arguments{
\item{inRaster}{RasterLayer (see \link[raster]{raster})}

\item{tiles}{a tile scheme generated using \code{\link{TileScheme}}}
}
\description{
Applies a tiling scheme to a RasterLayer and saves tiles to a temporary folder.
}
