% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TileHandlers.R
\name{TileApply}
\alias{TileApply}
\title{Tile Apply}
\usage{
TileApply(inList, maxCells, tileBuffer = 0, verbose = FALSE)
}
\description{
Takes a list of Raster objects and returns a tiling scheme.
  If the list contains multiple Raster objects, use TileDetector to detect existing tiles
  If the list contains a single small Raster object, return a single tile
  If the list contains a large Raster object (more cells than maxCells)
}

