% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tileDetector.R
\name{tileDetector}
\alias{tileDetector}
\title{Tile Detector}
\usage{
tileDetector(inRasters, reord = FALSE)
}
\arguments{
\item{inRasters}{a list of RasterLayers}

\item{reord}{logical. If set to \code{FALSE}, tiles will be stored in the same order in which
they appear in \code{inRasters}. If \code{TRUE}, tiles will be reordered by column and then by row.}
}
\description{
Function for detecting existing tiling scheme from a list of RasterLayers.
}
