% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TileHandlers.R
\name{TileInput}
\alias{TileInput}
\title{Tile Input}
\usage{
TileInput(input, inputName)
}
\description{
Takes either a single Raster object or paths to raster files and returns a list of Raster objects
}

