% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tileSave.R
\name{tileLoad}
\alias{tileLoad}
\title{Load Tile Scheme}
\usage{
tileLoad(filepath)
}
\arguments{
\item{filepath}{file path}
}
\description{
Load a Tile Scheme to a SHP file. The file needs to be originally saved using \code{tileSave}, since
some metadata (saved to an XML file) is required.
}
