% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tileSave.R
\name{tileSave}
\alias{tileSave}
\title{Save Tile Scheme}
\usage{
tileSave(ts, filepath, overwrite = FALSE)
}
\arguments{
\item{ts}{tileScheme}

\item{filepath}{file path}

\item{overwrite}{logical. Overwrite existing file}
}
\description{
Save a Tile Scheme to a SHP file
}
