% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tileScheme-methods.R
\name{identical}
\alias{identical}
\title{Identical}
\usage{
identical(x, y, num.eq = TRUE, single.NA = TRUE,
  attrib.as.set = TRUE, ignore.bytecode = TRUE,
  ignore.environment = FALSE, ignore.srcref = TRUE)
}
\description{
Determine if two 'tileScheme' objects are identical
}
