% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tileScheme-methods.R
\name{identical}
\alias{identical}
\alias{identical,tileScheme-method}
\title{Identical}
\usage{
identical(
  x,
  y,
  num.eq = TRUE,
  single.NA = TRUE,
  attrib.as.set = TRUE,
  ignore.bytecode = TRUE,
  ignore.environment = FALSE,
  ignore.srcref = TRUE
)

\S4method{identical}{tileScheme}(x, y)
}
\arguments{
\item{x, y}{'tileScheme' object}

\item{num.eq, single.NA, attrib.as.set, ignore.bytecode, ignore.environment, ignore.srcref}{arguments unused for 'tileScheme'}
}
\description{
Determine if two 'tileScheme' objects are identical
}
