% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tileScheme-methods.R
\name{length,tileScheme-method}
\alias{length,tileScheme-method}
\title{Length}
\usage{
\S4method{length}{tileScheme}(x)
}
\arguments{
\item{x}{'tileScheme' object}
}
\description{
Get the number of a cells contained in a 'tileScheme' object
}
