% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tileScheme-methods.R
\name{[,tileScheme,character,ANY-method}
\alias{[,tileScheme,character,ANY-method}
\alias{[,tileScheme,numeric,numeric-method}
\alias{[,tileScheme,numeric,missing-method}
\alias{[,tileScheme,missing,numeric-method}
\alias{[[,tileScheme,character,missing-method}
\title{Subset}
\usage{
\S4method{[}{tileScheme,character,ANY}(x, i, j, ..., drop = TRUE)

\S4method{[}{tileScheme,numeric,numeric}(x, i, j, ..., drop = TRUE)

\S4method{[}{tileScheme,numeric,missing}(x, i, j, ..., drop = TRUE)

\S4method{[}{tileScheme,missing,numeric}(x, i, j, ..., drop = TRUE)

\S4method{[[}{tileScheme,character,missing}(x, i, j, ..., exact = TRUE)
}
\arguments{
\item{x}{a 'tileScheme' object}

\item{i, j, ...}{indices specifying elements to extract}

\item{drop, exact}{arguments not used for 'tileScheme'}
}
\description{
Subset tiles using the single bracket operator. Subset geometry (tiles, buffs or nbuffs) using the double brackets
}
