% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tileDetector.R
\name{tileDetector}
\alias{tileDetector}
\title{Tile Detector}
\usage{
tileDetector(inRasters, roundCoords = 4)
}
\arguments{
\item{inRasters}{a list of RasterLayers}

\item{roundCoords}{numeric. Round input coordinates to this digit}
}
\description{
Function for detecting existing tiling scheme from a list of raster files or RasterLayers. This includes detecting buffers,
row and column positions, and re-ordering based on tile location.
}
